requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
sqlalchemy==2.0.27
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
alembic==1.13.1
redis==5.0.1
python-dotenv==1.0.1
pydantic[email]==2.6.1
email-validator==2.1.0.post1
bcrypt==4.1.2

# Testing dependencies
pytest==8.0.0
pytest-asyncio==0.23.5
pytest-cov==4.0.0
httpx==0.26.0
coverage==7.4.0 